fastapi-backend/requirements.txt
# FastAPI Backend Dependencies

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI API Integrations
google-generativeai==0.3.2
groq==0.11.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
hypothesis==6.98.3

# Utilities
python-dotenv==1.0.0
